@@ -287,7 +287,7 @@ class DataFilter extends Model
         if (!is_object($this->_searchModel) || $this->_searchModel instanceof \Closure) {
             $model = Yii::createObject($this->_searchModel);
             if (!$model instanceof Model) {
-                throw new InvalidConfigException('`' . get_class($this) . '::$searchModel` should be an instance of `' . Model::className() . '` or its DI compatible configuration.');
+                throw new InvalidConfigException('`' . get_class($this) . '::$searchModel` should be an instance of `' . Model::class . '` or its DI compatible configuration.');
             }
             $this->_searchModel = $model;
         }
@@ -301,7 +301,7 @@ class DataFilter extends Model
     public function setSearchModel($model)
     {
         if (is_object($model) && !$model instanceof Model && !$model instanceof \Closure) {
-            throw new InvalidConfigException('`' . get_class($this) . '::$searchModel` should be an instance of `' . Model::className() . '` or its DI compatible configuration.');
+            throw new InvalidConfigException('`' . get_class($this) . '::$searchModel` should be an instance of `' . Model::class . '` or its DI compatible configuration.');
         }
         $this->_searchModel = $model;
     }
