@@ -37,6 +37,6 @@ abstract class DbMutex extends Mutex
     public function init()
     {
         parent::init();
-        $this->db = Instance::ensure($this->db, Connection::className());
+        $this->db = Instance::ensure($this->db, Connection::class);
     }
 }
