@@ -233,7 +233,7 @@ class UrlRule extends CompositeUrlRule
                         Yii::debug([
                             'rule' => method_exists($rule, '__toString') ? $rule->__toString() : get_class($rule),
                             'match' => $result !== false,
-                            'parent' => self::className(),
+                            'parent' => self::class,
                         ], __METHOD__);
                     }
                     if ($result !== false) {
