@@ -60,7 +60,7 @@ abstract class CompositeUrlRule extends BaseObject implements UrlRuleInterface
                 Yii::debug([
                     'rule' => method_exists($rule, '__toString') ? $rule->__toString() : get_class($rule),
                     'match' => $result !== false,
-                    'parent' => self::className(),
+                    'parent' => self::class,
                 ], __METHOD__);
             }
             if ($result !== false) {
