@@ -253,7 +253,7 @@ abstract class BaseListView extends Widget
         }
         $pager = $this->pager;
         /** @var LinkPager $class */
-        $class = ArrayHelper::remove($pager, 'class', LinkPager::className());
+        $class = ArrayHelper::remove($pager, 'class', LinkPager::class);
         $pager['pagination'] = $pagination;
         $pager['view'] = $this->getView();
 
@@ -272,7 +272,7 @@ abstract class BaseListView extends Widget
         }
         $sorter = $this->sorter;
         /** @var LinkSorter $class */
-        $class = ArrayHelper::remove($sorter, 'class', LinkSorter::className());
+        $class = ArrayHelper::remove($sorter, 'class', LinkSorter::class);
         $sorter['sort'] = $sort;
         $sorter['view'] = $this->getView();
 
